pyproject.toml
[tool.poetry]
name = "college_statement"
version = "0.1.0"
description = ""
authors = ["Sashrik-Gupta <sashrik@infocusp.com>"]
readme = "README.md"
packages = [{ include = "Core" }]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"

fastmcp = ">=3.0.0b1"
langgraph = "^1.0.7"
langgraph-prebuilt = "^1.0.7"
langgraph-sdk = "^0.3.3"
langgraph-checkpoint-sqlite = "^3.0.3"
langsmith = "^0.6.8"
langchain-community = "^0.4.1"
langchain-core = "^1.2.8"
langchain-openai = "^1.1.7"
langchain-tavily = "^0.2.17"
langchain-groq = "^1.1.2"
wikipedia = "^1.4.0"
trustcall = "0.0.39"
pyngrok = "^7.5.0"
langgraph-cli = { version = "^0.4.12", extras = ["inmem"] }
kagglehub = "^0.4.2"
pandas = "^2.2.0"
matplotlib = "^3.10.8"
matplotlib-inline = "^0.2.1"
ruff = "^0.15.0"
langchain-google-vertexai = "^3.2.2"
langchain-google-genai = "^4.2.0"
amazon-product-review-scraper = "^0.9"
langchain-experimental = "^0.4.1"
e2b = "^2.13.2"
e2b-code-interpreter = "^2.4.1"
langchain-openrouter = "^0.0.2"
python-dotenv = "^1.2.1"
path = "^17.1.1"
fastapi = "^0.131.0"
streamlit = "^1.54.0"
pyaudio = "^0.2.14"
ipywidgets = "^8.1.8"
sounddevice = "^0.5.5"
numpy = "^2.4.2"



[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ipykernel (>=7.2.0,<8.0.0)"
]
